Add generic file splitting helper methods
Built off of this reference code:

    buf_left     = byte_count
    while buf_left.positive?
      cur_readsize = buf_left - FTP_CHUNKSIZE >= 0 ? FTP_CHUNKSIZE : buf_left
      buf = input_file.read(cur_readsize)
      break if buf == nil # rubocop:disable Style/NilComparison (from original)
      conn.write(buf)
      buf_left -= FTP_CHUNKSIZE
    end

From:

    https://github.com/ManageIQ/manageiq/pull/17549

This basically allows a shared form for uploading the source_input to a
file IO (whether that IO is a socket or or some other form of IO).

The method has been split up from the reference code since there are
some places where the loop itself might not be needed, but the code to
deliver a chunk at a time might be (see next commit for an example of
this using `MiqSwiftStorage`).

Also, since `read_single_chunk` is designed to be called on it's own, we
want to be able to eject with an empty string from that method itself.
This basically re-arranges some code (from the reference code) to allow
that, but maintain the same amount of conditional logic per chunk read
that was there previously.