Implements file spliting for MiqSwiftStorage#add
This code takes advantage of `fog-openstack`'s use of `Excon` as it's
HTTP client library, and it's implementation of `put_object` in
`Fog::Storage::OpenStack::Real`, which allows sending a block
(unfortunately, it doesn't pass the block from `files.create`, but we
won't talk about that...)

This takes advantage of the `read_single_chunk` method, defined in the
last commit, which will read at most a single DEFAULT_CHUNKSIZE per
block iteration, until it returns a empty string.

`read_single_chunk`'s role is to make sure that we only read up to the
`byte_count` from the `source_input`, and then return an empty string
when we have completed.