Fixes clearing tmpdir in generated_tmp_files.rb
The comment in the code should explain the issue for source_path.

For dest_path, this just adds auto clearing for the destination files as
well.  Requires some `defined?` checks since `dest_path` isn't something
that is defined as part of this helper, but it is used in basically all
of the spec files that use it for determining the destination file.