[FIXUP] Fix MiqSwiftStorage#container creation
Seems like some calls to Fog::Storage::OpenStack::Directories#get will
return `nil`, and not return an error.  This would cause errors down
the line.

So instead of investigating further, I just decided to create a new
method that is in charge of OpenStack container creation, and that is
called from `#container` if `nil` is returned, or the error for
`NotFound` is called to cover both cases.