[FIXUP][HACK] Require `pp`
Because we use the vmdb/logging, there is situations where the error
from there might need to be inspected in the logs, and we use
`.pretty_inspect` from 'pp' to do this.

Unfortunately, this doesn't seem to be a dependency of VMDB::Logger (or
associated libs), and this just happens to be loaded by other libraries.
This cause an error when running this code via the evm:db:* rake tasks
since `pp` is never required.

This is a HACK, for now, to basically make sure that dependency is
always available when this lib is loaded.  Later, we should really make
this a dependency of `VMDB::Logger`.