[FIXUP] Remove ::Settings based extra options
For me, this was causing errors when executing this via the `evm:db:*`
rake tasks (though worked fine in a `bin/rails c`) because Settings
wasn't loaded.

I took a peek at each of these, and most of these Settings are either
`nil` by default, or doesn't seem to affect posting to swift.  Removing
them for now seems like the best course of action.