[FIXUP] Use internal method for upload_single
Because of how `Fog::OpenStack` (and probably `Fog::Core`) is designed,
it has hidden the functionality to provide a block for streaming uploads
that is available out of the box with Excon.

Since there isn't a good way to do it with it's `File#create` method, I
opted to instantiate the object, and call the internal `Fog::Service`
object directly to make the request.  Most of what is normally done in
`Fog::Storage::OpenStack::Real.put_object` is the same as it was, and
this just allows properly setting the `request_block` parameter.

That said, there is a bug as well in `put_object`, since it will return
`202` status codes when doing a multipart upload, but only accepts 201
in it's parameters.  This updates it to accept both (which is also the
same expectations for `put_container` as well, so this isn't a terribly
crazy change).