[MiqSwiftStorage] Assign @container_name correctly
There was a bug when passing in a `:uri` to `MiqSwiftStorage` where if
the path for uri contained a filename as well:

    swift://example.com/container/path/to/file?query=params

It would assign the `@container_name` to the full path (minus the
leading slash), and not just the top level "directory".

This fixes that so it is correct.